Rename quarkus-junit5-vertx to quarkus-test-vertx and remove quarkus-junit5 dependency
