+++ integration-tests/hibernate-reactive-panache/src/test/java/io/quarkus/it/panache/reactive/TestReactiveTransactionTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import io.quarkus.hibernate.reactive.panache.Panache;
 import io.quarkus.test.TestReactiveTransaction;
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.vertx.UniAsserter;
+import io.quarkus.test.vertx.UniAsserter;
 
 @QuarkusTest
 public class TestReactiveTransactionTest {
