--- test-framework/junit5-vertx/src/main/resources/META-INF/services/io.quarkus.test.TestMethodInvoker
@@ -1,1 +1,0 @@
-io.quarkus.test.junit.vertx.RunOnVertxContextTestMethodInvoker
