+++ test-framework/pom.xml
@@ -29,7 +29,7 @@
         <module>junit5</module>
         <module>junit5-mockito</module>
         <module>junit5-mockito-config</module>
-        <module>junit5-vertx</module>
+        <module>vertx</module>
         <module>amazon-lambda</module>
         <module>arquillian</module>
         <module>devmode-test-utils</module>
