+++ test-framework/vertx/pom.xml
@@ -10,13 +10,13 @@
         <version>999-SNAPSHOT</version>
     </parent>
 
-    <artifactId>quarkus-junit5-vertx</artifactId>
-    <name>Quarkus - Test framework - JUnit 5 - Vert.x</name>
+    <artifactId>quarkus-test-vertx</artifactId>
+    <name>Quarkus - Test framework - Vert.x</name>
 
     <dependencies>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-junit5</artifactId>
+            <artifactId>quarkus-test-common</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
