+++ test-framework/vertx/src/main/java/io/quarkus/test/vertx/RunOnVertxContextTestMethodInvoker.java
@@ -1,4 +1,4 @@
-package io.quarkus.test.junit.vertx;
+package io.quarkus.test.vertx;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
