+++ test-framework/vertx/src/main/resources/META-INF/services/io.quarkus.test.TestMethodInvoker
@@ -1,0 +1,1 @@
+io.quarkus.test.vertx.RunOnVertxContextTestMethodInvoker
